Add app-vim/rust-mode.
I'm unsure whether the name rust-mode is right, but it installs not only
syntax/ script, so i think rust-syntax is not so good. I borrowed the
name from app-vim/python-mode.

I think I shouldn't fill the maintainer field with the your address even
if it seems natural, so I don't write metadata.xml. If you are willing
to maintain this, could you write it?

Signed-off-by: OGINO Masanori <masanori.ogino@gmail.com>