dev-lang/rust: set sub-slot
While everything built with Rust is currently static there is a
possibility that at some point in the future we might build things in
the tree or in the overlay dynamically. Upstream has said they don't
guarantee the ABI between rlibs built with different MAJOR.MINOR's of
the compiler so those would need to be rebuilt when that changes. The
only way to reliably do that is to set a sub-slot and have those
packages include dev-lang/rust:= as their dependency.