dev-lang/rust: improve stage0 fetching
The stage0 tarball is specified in parts in src/snapshots.txt, if you
build rust and you don't have the most recent version the build system
will pull it down when you start compiling. This violates Gentoo's
policy of not doing a fetch outside of SRC_URI or src_fetch() and for
users with FEATURES=network-sandbox it will not work. So it is important
we have the right version so break out all the fields near the top so
people that bump this package hopefully pay attention and keep it up to
date. The previous behavior also copied in any stage0 images that might
live in your distfiles directory rather than just the one we need.