media-gfx/mozjpeg: new ebuild
This ebuild is an horrible hack to be able to install libjpeg-turbo and
mozjpeg side by side.
The elf executables are moved in libexec and a wrapper for them created
in /usr/bin.
The wrapper preload the mozjpeg library which replace libjpeg.so
_all_ the executables are prefixed by the string "moz" i.e. cjpeg
becomes mozcjpeg.

The project readme file states:

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Mozilla JPEG Encoder Project

This project's goal is to reduce the size of JPEG files without reducing quality or compatibility with the vast majority of the world's deployed decoders.

The idea is to reduce transfer times for JPEGs on the Web, thus reducing page load times.

'mozjpeg' is not intended to be a general JPEG library replacement. It makes tradeoffs that are intended to benefit Web use cases and focuses solely on improving encoding. It is best used as part of a Web encoding workflow. For a general JPEG library (e.g. your system libjpeg), especially if you care about decoding, we recommend libjpeg-turbo.
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~