dev-lang/python: add "pgo" to IUSE
Python can now be built with --enable-optimizations.
This option at the moment enable profile guided optimization for gcc and
clang compilers.
It seem that link time optimizations are left as an exercise for the
user (which is a good thing)

The build phase is modified so that the sourcecode is compiled with
"-fprofile-generate" flag first (for gcc).
Then some tests are run, exercising the binaries and creating profiling
data.
Then the final executable is built with "-fprofile-use
-fprofile-correction" flags taking advantage of the gathered data.